nenzfr/estcj: numerous small but important changes and bug fixes
estcj.py: wouldn't run for the 'stn' and 'stnp' options. The cause was
          the "secant" function feeding a negative value for p_on_p0 to
          "expand_from_stagnation". I have now limited p_on_p0 to the
          range [1.0/p0, 1.0] in the "secant" call.
cea-cases/: I have added air-no-ions, N2-no-ions and air5species-no-ions
          cases and adjusted the air5species case accordingly.
nenzfr.py/nozzle.input.template: we now use the no-ions cases/LUT for
          an equilibrium gas
run_template_Barrine.sh: bug fix. We needed to change into the working
          directory and specify the e3bin path in order to be able to
          call nenzfr

+++ app/nenzfr/nenzfr_quasi_transient.py
@@ -121,7 +121,7 @@ def main():
     op.add_option('--tstart', dest='tstart', type='float', default='1.5e-3',
                   help=("time at which the first slice of the input pressure "
                         "trace is to begin [default: %default]"))
-    op.add_option('--nsteps', dest='nsteps', type='float', default=5,
+    op.add_option('--nsteps', dest='nsteps', type='int', default=5,
                   help=("number of slices to use [default: %default]"))
     op.add_option('--dt', dest='dt', type='float', default=0.5e-3,
                   help=("width of each averaging slice [default: %default]"))
